+++ crates/samples/create_window_sys/src/main.rs
@@ -35,7 +35,7 @@ fn main() {
 
 extern "system" fn wndproc(window: HWND, message: u32, wparam: WPARAM, lparam: LPARAM) -> LRESULT {
     unsafe {
-        match message as u32 {
+        match message {
             WM_PAINT => {
                 println!("WM_PAINT");
                 ValidateRect(window, std::ptr::null());
